Make helpers for filter population
This is another boilerplate reduction for new filters based on new fields. All but enum types are no a single method call for setting, including long and long? values that used to take up like 20 lines apiece. This should reduce the amount of possible issues in setting up the filters.