Use "include" instead of "skip" for D2D
This is a semi-breaking change. The normal functionality of not specifying any hashes will default to CRC32, MD5, and SHA-1 as expected. However, all other flags have been renamed and act as an include (like how the SHA-256 and above functioned previously). If any of those flags are specified, then it overrides the default behavior and ONLY uses the hashes specified. So to get the same behavior as `--skip-md5`, a user would do `--include-crc --include-sha1`.